requirements.txt
flask==2.0.1
anthropic==0.18.0
python-dotenv==0.19.0
gunicorn==20.1.0
werkzeug==2.0.3
click>=7.0
itsdangerous>=2.0
Jinja2>=3.0
MarkupSafe>=2.0
